Fix in IncreaseEmergencyPlace

+++ Engine/src/it/uniba/hazard/engine/cards/IncreaseEmergencyPlace.java
@@ -36,7 +36,7 @@ public class IncreaseEmergencyPlace extends EventCard{
         listLocation.toArray(l);
 
         while(true){
-            int randomIndex = new Random().nextInt()*l.length-1;
+            int randomIndex = (int) ((new Random().nextDouble())*(l.length-1));
             //verifico che ci sia l'emergenza in quella location
             if(l[randomIndex].getEmergencyLevel(randomEmergency) > 0) {
                 int currentEmergencyLevel = l[randomIndex].getEmergencyLevel(randomEmergency);
